+++ changelog.md
@@ -9,6 +9,7 @@
 - Favorite events that have passed do not show in the list
 - Fix to correctly choose todays day if its available in events
 - Search now has a "go" button
+- Map of Favorite Art, Events and Camps
 - Show events that are happening now
 - Better message when no events match
 - Favorites can be searched